Refactor input reading to a helpers package
